LPD-19409 Move
