+++ modules/apps/object/object-service/src/main/java/com/liferay/object/internal/sort/ObjectEntry1ToMRelationshipSortDSLQueryVisitor.java
@@ -5,7 +5,6 @@
 
 package com.liferay.object.internal.sort;
 
-import com.liferay.object.internal.sort.relationship.RelationshipSort;
 import com.liferay.object.model.ObjectDefinition;
 import com.liferay.object.model.ObjectRelationship;
 import com.liferay.object.petra.sql.dsl.DynamicObjectDefinitionTable;
