+++ modules/apps/object/object-service/src/main/java/com/liferay/object/internal/sort/RelationshipSort.java
@@ -3,9 +3,8 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.object.internal.sort.relationship;
+package com.liferay.object.internal.sort;
 
-import com.liferay.object.internal.sort.Sort;
 import com.liferay.object.model.ObjectDefinition;
 import com.liferay.object.model.ObjectRelationship;
