+++ modules/apps/object/object-service/src/main/java/com/liferay/object/internal/sort/SortDSLQueryBuilderContributor.java
@@ -8,7 +8,6 @@ package com.liferay.object.internal.sort;
 import com.liferay.object.constants.ObjectRelationshipConstants;
 import com.liferay.object.internal.sort.field.ObjectEntryFieldSortDSLQueryBuilderContributor;
 import com.liferay.object.internal.sort.relationship.ObjectEntry1ToMRelationshipSortDSLQueryBuilderContributor;
-import com.liferay.object.internal.sort.relationship.RelationshipSort;
 import com.liferay.object.model.ObjectDefinition;
 import com.liferay.object.model.ObjectRelationship;
 import com.liferay.object.petra.sql.dsl.DSLQueryBuilder;
