+++ modules/apps/object/object-service/src/main/java/com/liferay/object/internal/sort/SortDSLQueryVisitor.java
@@ -6,7 +6,6 @@
 package com.liferay.object.internal.sort;
 
 import com.liferay.object.constants.ObjectRelationshipConstants;
-import com.liferay.object.internal.sort.relationship.RelationshipSort;
 import com.liferay.object.model.ObjectDefinition;
 import com.liferay.object.model.ObjectRelationship;
 import com.liferay.object.relationship.util.ObjectRelationshipUtil;
