+++ modules/apps/object/object-service/src/main/java/com/liferay/object/internal/sort/relationship/ObjectEntry1ToMRelationshipSortDSLQueryBuilderContributor.java
@@ -6,6 +6,7 @@
 package com.liferay.object.internal.sort.relationship;
 
 import com.liferay.object.internal.sort.BaseSortDSLQueryBuilderContributor;
+import com.liferay.object.internal.sort.RelationshipSort;
 import com.liferay.object.internal.sort.Sort;
 import com.liferay.object.model.ObjectDefinition;
 import com.liferay.object.model.ObjectRelationship;
